+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Information section in media server settings for some additional details on initial setup
 - Descriptions for Radarr, Sonarr, Seerr, Tautulli, MDBlist, and OMDb service configurations
+- Additional information for metadata tasks
 
 ### Changed